big update: change database to superbase

+++ client/src/components/CompletedTab.tsx
@@ -5,14 +5,17 @@ import { Trash } from "lucide-react";
 
 type TodoType = {
   id: number;
-  title: string;
+  user_id: string;
+  task: string;
   note: string;
-  pinned: boolean;
-  completed: boolean;
-  remind_date: Date;
+  priority: "none" | "low" | "medium" | "high";
+  location?: string;
+  attachment?: string;
+  is_complete: boolean;
+  is_pin: boolean;
   created_at: Date;
   updated_at: Date;
-  deleted_at: Date;
+  remind_at: Date;
 };
 
 type Props = {
@@ -27,16 +30,19 @@ const CompletedTab = ({ todos, setTodos }: Props) => {
         <h1 className="font-bold text-lg">Completed</h1>
         <button
           className="flex gap-[6px] w-20 h-7 items-center hover:bg-red-400/10 px-2 rounded-full hover:text-red-500 transition-all"
-          onClick={() => {
-            {
-              todos
-                .filter((todo) => todo.completed)
-                .forEach(async (todo) => {
-                  setTodos(todos.filter((todo) => !todo.completed));
-                  await axios.delete(
-                    `https://todo-app-avvn.onrender.com/todos/${todo.id}`
-                  );
-                });
+          onClick={async () => {
+            const completedTodos = todos.filter((todo) => todo.is_complete);
+
+            if (completedTodos.length > 0) {
+              setTodos((prevTodos) =>
+                prevTodos.filter((todo) => !todo.is_complete)
+              );
+
+              for (const todo of completedTodos) {
+                await axios.delete(
+                  `https://todo-app-avvn.onrender.com/todos/${todo.id}`
+                );
+              }
             }
           }}
         >
@@ -48,7 +54,7 @@ const CompletedTab = ({ todos, setTodos }: Props) => {
       </div>
       <div className="flex flex-col gap-2 overflow-x-clip overflow-y-auto px-2 ">
         {todos
-          .filter((todo) => todo.completed)
+          .filter((todo) => todo.is_complete)
           .map((todo) => (
             <Todo key={todo.id} todo={todo} todos={todos} setTodos={setTodos} />
           ))}
